+++ lib/eloquent.dart
@@ -5,6 +5,7 @@ export 'src/query/query_builder.dart';
 export 'src/query/sql_constants.dart';
 
 export 'src/query/join_clause.dart';
+export 'src/query/on_conflict_clause.dart';
 
 export 'src/grammar.dart'; //BaseGrammar
 export 'src/query/grammars/query_grammar.dart';